feat(cabinet): tenant impersonation flow and auto cabinet user bootstrap

+++ app/Filament/Widgets/RecentTenantRequestsWidget.php
@@ -132,7 +132,7 @@ class RecentTenantRequestsWidget extends BaseTableWidget
                     $created = CarbonImmutable::parse($record->created_at)->setTimezone($tz)->startOfDay();
 
                     // Положительное число дней, сколько заявка "в работе"
-                    return max($created->diffInDays($today), 0);
+                    return (int) max((int) $created->diffInDays($today), 0);
                 }),
         ];
     }
@@ -180,3 +180,4 @@ class RecentTenantRequestsWidget extends BaseTableWidget
         return $tz;
     }
 }
+
